Refactor the transactions out of the callback
