+++ src/types.ts
@@ -1,7 +1,6 @@
-import { either as E } from 'fp-ts'
 import * as t from 'io-ts'
 
-export const signMessageType = t.strict({
+export const SignMessage = t.strict({
   id: t.string,
   tag: t.string,
   payload: t.strict({
@@ -12,9 +11,9 @@ export const signMessageType = t.strict({
   }),
 })
 
-export type signMessageType = t.TypeOf<typeof signMessageType>
+export type SignMessage = t.TypeOf<typeof SignMessage>
 
-export const signMessageErrorType = t.strict({
+export const SignMessageError = t.strict({
   id: t.string,
   tag: t.string,
   payload: t.strict({
@@ -28,4 +27,10 @@ export const signMessageErrorType = t.strict({
   }),
 })
 
-export type signMessageErrorType = t.TypeOf<typeof signMessageErrorType>
+export type SignMessageError = t.TypeOf<typeof SignMessageError>
+
+export type DecodedIdToken = {
+  payload: any
+  header?: any
+  signature?: string
+}
